flex: build vertices map in normalized coordinates
I thought using pixel coordinates would help applying width/height constraints
on geometry change, but I decided to not allowing changing the grid relation
in that case and therefore no need to work around the pixel resolution
problems.

It can be flaky to use floating point number as map keys, but in our use case,
the original keys should be available and the map can be looked up by them.